+++ src/App.tsx
@@ -1,15 +1,15 @@
-import { AuthContainer } from './hooks/auth';
 import { Outlet } from 'react-router-dom';
 import Header from './components/Header';
+import { Fragment } from 'react';
 
 // TODO: Prevent NOT logged in user from Accessing /books page (through changing browser url)
 
 const App = (): JSX.Element => {
   return (
-    <AuthContainer.Provider>
+    <Fragment>
       <Header />
       <Outlet />
-    </AuthContainer.Provider>
+    </Fragment>
   );
 }
